- Modified Makefile to compile files for debug - Created double_fault_handler and generic_handler functions in handler_functions.c

+++ include/pic/pic.h
@@ -2,7 +2,7 @@
 #define __8259_PIC__
 #include <stdint.h>
 
-void PIC_sendEOI(uint8_t irq);
+void PIC_sendEOI(uint32_t irq);
 void PIC_remap(uint32_t offset1, uint32_t offset2);
 uint16_t __pic_get_irq_reg(int ocw3);
 uint16_t pic_get_irr(void);
